!!BREAKING CHANGE!! Prototype treating DateTime/Decimal/Guid/TimeSpan as sub-message; manually unrolled serialization (for peformance), but uses regular sub-message deserialization; added Char/TimeSpan support
git-svn-id: http://protobuf-net.googlecode.com/svn/trunk@116 b72047a4-3652-0410-9617-b3994939e97b

+++ protobuf-net/Serializers/ZigZagSerializer_Int32.cs
@@ -5,7 +5,7 @@ namespace ProtoBuf
     {
         string ISerializer<int>.DefinedType { get { return ProtoFormat.SINT32; } }
 
-        private static long ZigInt32(int value)
+        private static int ZigInt32(int value)
         {
             return ((value << 1) ^ (value >> 31));
         }
